(fix) Update files to reference the newly renamed server files, so that grunt deploy works.
