+++ Gruntfile.js
@@ -59,7 +59,7 @@ module.exports = function(grunt) {
 
     nodemon: {
       dev: {
-        script: 'index.js'
+        script: 'server.js'
       }
     },
